+++ tests/conftest.py
@@ -7,26 +7,4 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-import logging
-
-import pytest
-
-from .executor import spawn_gateway
-
 pytest_plugins = "anemoi.utils.testing"
-logger = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="session")
-def shared_gateway():
-    """Shared gateway process and URL for all tests to avoid ZMQ conflicts."""
-    url, process = spawn_gateway(max_jobs=4)
-    try:
-        yield url
-    except Exception as e:
-        # NOTE we log like this in case shm shutdown freezes
-        logger.exception(f"gotten {repr(e)}, proceed with shm shutdown")
-        raise
-    finally:
-        if process and process.is_alive():
-            process.kill()
